drivers: gpu: Fix -Wpointer-to-int-cast
/home/xiaolegun/lineage/kernel/xiaomi/mt6885/drivers/gpu/drm/mediatek/mtk_drm_crtc.c:2529:
15: error: cast to smaller integer type 'unsigned int' from 'struct cmdq_pkt *' [-Werror,
-Wpointer-to-int-cast]
                                __func__, (unsigned int)*cmdq_handle);

+++ drivers/gpu/drm/mediatek/mtk_drm_crtc.c
@@ -2526,7 +2526,7 @@ void mtk_crtc_pkt_create(struct cmdq_pkt **cmdq_handle, struct drm_crtc *crtc,
 	*cmdq_handle = cmdq_pkt_create(cl);
 	if (IS_ERR_OR_NULL(*cmdq_handle)) {
 		DDPPR_ERR("%s create handle fail, %x\n",
-				__func__, (unsigned int)*cmdq_handle);
+				__func__, *cmdq_handle);
 		return;
 	}
